requirements.txt
aiogram==3.4.1
telethon==1.33.1
python-dotenv==1.0.0
